requirements.txt
streamlit==1.51.0
python-dotenv==1.0.1
openai==1.52.2
Pillow==11.0.0
